Refactor code and add new endpoint for robot movement
The main changes involve renaming 'Position' and 'MovementCommand' to 'Position' and 'Movement' respectively and moving them to a new package. This also changes corresponding function parameters from the old to the new names. Additionally, this commit introduces a new endpoint '/moves' that will calculate the changes and return the resulting movements for an array of positions.